MAJOR - Refactored server.js to seperate concerns
