+++ server/startup/config.js
@@ -1,0 +1,7 @@
+const config = require('config');
+
+module.exports = function() {
+    if (!config.get('jwtPrivateKey')) {
+        throw new Error('FATAL ERROR: jwtPrivateKey is not defined...');
+    };
+}
