+++ server/startup/routes.js
@@ -1,0 +1,33 @@
+const bodyParser = require('body-parser');
+
+// Error handling and Logging
+const error = require('../middleware/winston'); // middleware/error.js which is running [npm] winston for error handling
+
+// Routes - Using Express
+const auth = require('../routes/api/auth');
+const interceptor = require('../routes/api/interceptor');
+const team = require('../routes/api/team');
+const intercept = require('../routes/api/intercept');
+const zones = require('../routes/api/zones');
+const country = require('../routes/api/country');
+const users = require('../routes/users');
+const news = require('../routes/api/news');
+const logs = require('../routes/api/log');
+
+module.exports = function(app) {
+    // Bodyparser Middleware
+    app.use(bodyParser.json());
+
+    // Express Routes - Endpoints to connect to through the browser. (Housed routes folder)
+    app.use('/api/auth', auth);
+    app.use('/api/interceptor', interceptor); // Route for manipulating interceptors
+    app.use('/api/team', team); // Route for Teams
+    app.use('/api/intercept', intercept); // Route for triggering an interception
+    app.use('/api/zones', zones); // Route for inputing zones
+    app.use('/api/country', country); // Route for inputing countries
+    app.use('/user', users); // Route for dealing with Users
+    app.use('/api/news', news); // Route for the news desks
+    app.use('/api/logs', logs); // Route for logs
+
+    app.use(error.routeError);
+}
